+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file. See [Keep a
 
 ## Unreleased
 
+### Added
+
+- GDT and GTM (`D` and `T`) now work in compiled programs.
+
 ## [1.6.2] - 2023-11-21
 
 ### Fixed